Change `/local` to `<local>`
Rebol2's "hack" to implement local variables was to make them arguments
to a refinement.  This had several downsides, including the loss of
/LOCAL as a meaningful refinement for user use (e.g. TIME/LOCAL).
But also, it meant callers were able to "inject" state for local
variables of functions--which wasn't a great idea.

Ren-C added `<local>` to give true locals, and reclaim /LOCAL as an
ordinary tag.  But code using `/LOCAL` would still work...up until
now.  Refinements arguments *are* the refinement, so /LOCAL could be
at most one value.

This switches to using the tag.