Update PARSE conventions for ANY, HERE, SEEK
In order to use the old conventions (e.g. plain SET-WORD! for marking, and plain
GET-WORD! for seeking) this would have to use PARSE2.

PARSE2 will still be supported (e.g. Redbol) but the conventions that are being
adopted into PARSE from UPARSE are really the way forward.

This changes to use WHILE instead of ANY (as UPARSE ANY is not an iterative
construct, and WHILE makes more sense as being iterative).  It also uses the
SET-WORD! form for marking which involves using the TAG! combinator
of `<here>`.  It is chosen to be a tag precisely for cases like this script, which
want to be able to call a variable HERE.