Use .r as indicator of reviewed/updated files
The historical file extension of .r is being restored, vs. using .reb,
due to less noise and less contention for the filename in syntax
highlighters in the future due to the AI era.

Rather than blindly change all the files, this uses the extension
change as a way of tracking which files have been reviewed for the
modern interpreter.