Switch back to OPT for optional rules
The concept of using TRY is actually not dead, it just depends on
if you want a void or a null.

    parse "a" ["a" var: opt "b"]

That will unset var.  However, TRY will set it to null:

    parse "a" ["a" var: try "b"]

It only matters if you're using the expression result.  This changes back to OPT just to be closest to the original source.