Type constraints as terminal slash (not leading &)
The idea of TYPE-XXX! as types with leading ampersands fell
from favor, with datatypes becoming antiforms.

  https://rebol.metaeducation.com/t/old-idea-revisited-datatype-as-antiform/2324

Type constraints are functions, and you use them for matching in
the parse dialect with terminal slashes.

Things like GET-WORD! and SET-WORD! are CHAIN! and no longer
fundamental types.  This permits things like `(foo):` to exist, as the
structure is generalized...without creating a combinatoric explosion
of things like SET-GROUP!s or GET-INTEGER!s etc.