BINARY! => BLOB! rename
The word "binary" is very overloaded...not just to mean literally
base two (1 and 0) but also for naming EXE and OBJ files, and in
general is a fairly overloaded term.

There was also the issue of inside the source code, differentiating
between an allocated range of bytes, and the actual type which
carries and index and acts as a series.  Choosing the shorter and
more widespread name "BLOB!" helps these issues across the board,
while letting code be more brief at the same time.