True/False/Null changes
The idea of the system has centralized on a single falsey state
which is the null antiform.  The lone underscore represents the
space character literal--and while it can be used in various places
to indicate "blankness" it's really just a space.  The reified form
of a null is the quasi-word ~null~, which can be degraded or
reified when quasiforms are out of band... or fully lifted and
unlifted when they are not.  There are no longer `#[true]` or
`#[false]` forms, just the words TRUE and FALSE which you can
turn into antiform logic with TRUE? or FALSE? questions, and
do the reverse construction to words with BOOLEAN.  If you do
choose to use words for logic, it tends to make sense to use the
words that make the most sense, like YES and NO:

  https://rebol.metaeducation.com/t/flexible-logic-system-terminology/2252