SETIFY instead of TO SET-WORD!
SET-WORD! is not a fundamental type, it is now a CHAIN! with
a space in the second slot.

This has benefits, e.g. `(foo):` is a CHAIN! too, permitting you to
have other elements in that spot.  But it does mean that you have
to use destructuring and pattern matching.

So instead of `to set-word! key` you could say:

     compose '(key):

But there is also a SETIFY operation that does this more efficiently:

    setify key